genai_nike/requirements.txt
fastapi
openai
httpx
python-dotenv
pillow
